src_docker/admanagerworkflow/requirements.txt
pytz==2023.3
pyodbc
pre-commit
google-cloud-bigquery
python-dotenv
pyyaml
boto3
pandas
pandas-gbq
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
google-api-python-client
googleads
requests
gspread
oauth2client
pyjwt
cryptography
pydantic
pydantic[email]
sqlalchemy
dateparser
pytest
pyyaml
pytest-cov
pylint
flake8
chardet
s3fs
tenacity
botocore
greenlet
dateparser
db-dtypes
dill
dnspython
email-validator
numpy
pyarrow
python-dotenv
backoff
